+++ src/Components/AppConfig/DefaultColumns/Config.tsx
@@ -27,8 +27,10 @@ const Config = () => {
   return (
     <main className={styles.main}>
       <div className={styles.introText}>
-        <Badge color={'blue'} text={'Experimental'} />
-        <span>Configure default fields to display instead of the full log line:</span>
+        <Badge color={'blue'} text={'Beta'} />
+        <span>
+          Configure the fields to show by default. These can replace the full log line or be displayed next to it.
+        </span>
       </div>
 
       <ErrorBoundaryAlert>