+++ src/Components/Table/LogLineCellComponent.tsx
@@ -8,7 +8,6 @@ import { CustomCellRendererProps, useTheme2 } from '@grafana/ui';
 
 import { getBodyName } from '../../services/logsFrame';
 import { DETECTED_LEVEL } from './constants';
-import { OTEL_LOG_LINE_ATTRIBUTES_FIELD_NAME } from 'Components/ServiceScene/LogOptionsScene';
 import { useQueryContext } from 'Components/Table/Context/QueryContext';
 import { LogLineState, useTableColumnContext } from 'Components/Table/Context/TableColumnsContext';
 import { DefaultCellWrapComponent } from 'Components/Table/DefaultCellWrapComponent';
@@ -16,6 +15,7 @@ import { LineActionIcons } from 'Components/Table/LineActionIcons';
 import { LogLinePill } from 'Components/Table/LogLinePill';
 import { RawLogLineText } from 'Components/Table/RawLogLineText';
 import { Scroller } from 'Components/Table/Scroller';
+import { OTEL_LOG_LINE_ATTRIBUTES_FIELD_NAME } from 'services/logFieldNames';
 
 export type SelectedTableRow = {
   id: string;