+++ src/module.tsx
@@ -3,6 +3,7 @@ import { lazy } from 'react';
 import { AppPlugin } from '@grafana/data';
 import { initPluginTranslations } from '@grafana/i18n';
 
+import type { JsonData } from './Components/AppConfig/AppConfig';
 import pluginJson from 'plugin.json';
 import {
   SuspendedEmbeddedLogsExploration,
@@ -50,13 +51,24 @@ const DefaultColumnsConfig = lazy(async () => {
   return await import('./Components/AppConfig/DefaultColumns/Config');
 });
 
-export const plugin = new AppPlugin<{}>()
+const ServiceSelectionConfig = lazy(async () => {
+  await initPluginTranslations(pluginJson.id);
+  return await import('./Components/AppConfig/ServiceSelection/Config');
+});
+
+export const plugin = new AppPlugin<JsonData>()
   .setRootPage(App)
   .addConfigPage({
     body: AppConfig,
     icon: 'cog',
     id: 'configuration',
     title: 'Configuration',
+  })
+  .addConfigPage({
+    body: ServiceSelectionConfig,
+    icon: 'home-alt',
+    id: 'admin-service-selection',
+    title: 'Landing Page',
   })
   .addConfigPage({
     body: DefaultColumnsConfig,